projects/stock-project/requirements.txt
pandas
yfinance
psycopg2-binary
python-dotenv
tenacity
sqlalchemy